Use enum class for ball type

+++ core/source/editor/PlayTestScreen.cpp
@@ -64,5 +64,5 @@ void PlayTestScreen::render() {
 
 void PlayTestScreen::doResize() {
 	uiManager.resize(width, height, dpiScale);
-	game.resize(width, height);
+	game.resize((float)width, (float)height);
 }
